+++ PingIt/Features/Authentication/Models/AuthRoute.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+enum AuthRoute: Hashable {
+    case login
+    case register
+    case forgotPassword
+    case termsOfService
+}